+++ integration-tests/call-runtime/lib.rs
@@ -165,7 +165,7 @@ mod runtime_call {
                 .submit()
                 .await
                 .expect("instantiate failed");
-            let mut call = contract.call::<RuntimeCaller>();
+            let mut call_builder = contract.call_builder::<RuntimeCaller>();
 
             let receiver: AccountId = default_accounts::<DefaultEnvironment>().bob;
 
@@ -180,7 +180,7 @@ mod runtime_call {
 
             // when
             let transfer_message =
-                call.transfer_through_runtime(receiver, TRANSFER_VALUE);
+                call_builder.transfer_through_runtime(receiver, TRANSFER_VALUE);
 
             let call_res = client
                 .call(&ink_e2e::alice(), &transfer_message)
@@ -229,13 +229,13 @@ mod runtime_call {
                 .submit()
                 .await
                 .expect("instantiate failed");
-            let mut call = contract.call::<RuntimeCaller>();
+            let mut call_builder = contract.call_builder::<RuntimeCaller>();
 
             let receiver: AccountId = default_accounts::<DefaultEnvironment>().bob;
 
             // when
-            let transfer_message =
-                call.transfer_through_runtime(receiver, INSUFFICIENT_TRANSFER_VALUE);
+            let transfer_message = call_builder
+                .transfer_through_runtime(receiver, INSUFFICIENT_TRANSFER_VALUE);
 
             let call_res = client
                 .call(&ink_e2e::alice(), &transfer_message)
@@ -265,10 +265,10 @@ mod runtime_call {
                 .submit()
                 .await
                 .expect("instantiate failed");
-            let mut call = contract.call::<RuntimeCaller>();
+            let mut call_builder = contract.call_builder::<RuntimeCaller>();
 
             // when
-            let transfer_message = call.call_nonexistent_extrinsic();
+            let transfer_message = call_builder.call_nonexistent_extrinsic();
 
             let call_res = client
                 .call(&ink_e2e::alice(), &transfer_message)