+++ integration-tests/upgradeable-contracts/delegator/lib.rs
@@ -108,12 +108,12 @@ pub mod delegator {
                 .await;
 
             let mut constructor = DelegatorRef::new_default();
-            let call_builder = client
+            let contract = client
                 .instantiate("delegator", &origin, &mut constructor)
                 .submit()
                 .await
                 .expect("instantiate failed");
-            let mut call_builder_call = call_builder.call::<Delegator>();
+            let mut call_builder = contract.call_builder::<Delegator>();
 
             let code_hash = client
                 .upload("delegatee", &origin)
@@ -123,7 +123,7 @@ pub mod delegator {
                 .code_hash;
 
             // when
-            let call_delegate = call_builder_call.inc_delegate(code_hash);
+            let call_delegate = call_builder.inc_delegate(code_hash);
 
             let result = client.call(&origin, &call_delegate).submit().await;
             assert!(result.is_ok(), "delegate call failed.");
@@ -133,9 +133,9 @@ pub mod delegator {
 
             // then
             let expected_value = 4;
-            let call = call_builder.call::<Delegator>();
+            let call_builder = contract.call_builder::<Delegator>();
 
-            let call_get = call.get_counter();
+            let call_get = call_builder.get_counter();
             let call_get_result = client
                 .call(&origin, &call_get)
                 .dry_run()
@@ -161,12 +161,12 @@ pub mod delegator {
 
             // given
             let mut constructor = DelegatorRef::new(10);
-            let call_builder = client
+            let contract = client
                 .instantiate("delegator", &origin, &mut constructor)
                 .submit()
                 .await
                 .expect("instantiate failed");
-            let mut call_builder_call = call_builder.call::<Delegator>();
+            let mut call_builder = contract.call_builder::<Delegator>();
 
             let code_hash = client
                 .upload("delegatee", &origin)
@@ -176,7 +176,7 @@ pub mod delegator {
                 .code_hash;
 
             // when
-            let call_delegate = call_builder_call.add_entry_delegate(code_hash);
+            let call_delegate = call_builder.add_entry_delegate(code_hash);
             let result = client.call(&origin, &call_delegate).submit().await;
             assert!(result.is_ok(), "delegate call failed.");
 
@@ -188,7 +188,7 @@ pub mod delegator {
             // Alice's address
             let address = AccountId::from(origin.public_key().to_account_id().0);
 
-            let call_get_value = call_builder_call.get_value(address);
+            let call_get_value = call_builder.get_value(address);
             let call_get_result = client
                 .call(&origin, &call_get_value)
                 .submit()